backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
numpy==1.24.2
scipy==1.10.1
feedparser==6.0.10
google-generativeai==0.2.2
